+++ src/ChaiLove.cpp
@@ -90,19 +90,11 @@ bool ChaiLove::update() {
 				event.quit();
 				return !event.quitstatus;
 				break;
-			case SDL_MOUSEMOTION:
-				mouse.moveEvent(sdlEvent.motion);
-				break;
-			case SDL_MOUSEBUTTONDOWN:
-				mouse.buttonEvent(sdlEvent.button);
-				break;
-			case SDL_MOUSEBUTTONUP:
-				mouse.buttonEvent(sdlEvent.button);
-				break;
 		}
 	}
 
 	// Update some of the sub-systems.
+	mouse.update();
 	joystick.update();
 	keyboard.update();