+++ CHANGELOG.md
@@ -7,10 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## 0.14.3 - Unreleased
 ### Changed
 - Updated the input description names
-- Switch Keyboard input from SDL API to libretro API
+- Switch Keyboard/Mouse input from SDL API to libretro API
 
 ### Added
 - `love.keyboard.isScancodeDown()`
+- `love.mouse` `x1` and `x2` buttons
 
 ## 0.14.2 - 2018-02-17
 ### Changed